+++ examples/digital_twin.py
@@ -29,9 +29,9 @@ with mjtb.Wrapper(urdf,
                   meshdir=meshes,
                   initial_conditions=initial,
                   controller=real_time) as ur5:
-    ur5.liveView(show_menu=False) # Open the simulation window
+    ur5.launch(show_menu=False) # Open the simulation window
     ur5.gravity = [0, 0, 0]
 
     for qpos in cycle(qpos_sequence):
-        ur5.controller(ur5._model, ur5._data, {"qpos": qpos})
+        ur5.controller(ur5.model, ur5.data, {"qpos": qpos})
         time.sleep(5.0 / num_steps)