+++ mujoco_toolbox/__init__.py
@@ -1,31 +1,31 @@
-"""Mujoco Toolbox.
+"""Mujoco Toolbox
 ==============
 
-A toolbox for working with MuJoCo simulations. This package provides various
-utilities and controllers to facilitate a faster simulation process.
+A toolbox for working with MuJoCo simulations.
+
+This package provides various utilities and controllers to facilitate a faster 
+simulation process.
 
 Modules
 --------
-- `Wrapper`: Contains the Wrapper class for interfacing with MuJoCo.
-- `Controller`: Includes sineController, cosineController, and randomController
-    for controlling simulations.
-- `Builder`: Contains the Builder class for creating and merging MuJoCo models.
+- `wrapper`: Contains the Wrapper class for interfacing with MuJoCo.
+- `controllers`: Includes pre-built functions for controlling simulations.
+- `builder`: Contains the Builder class for creating and merging MuJoCo models.
 - `assets`: Contains pre-defined assets for building MuJoCo models.
 
 Constants
 ----------
 - CAPTURE_PARAMETERS: List of MjData fields to capture during simulation.
 
-License
---------
-This project is licensed under the `MIT License`. See the `LICENSE` file for details.
+This project is licensed under the `MIT License`. See the `LICENSE` file for 
+details.
 
 Notes
 -----
 This package is still under development. Report any issues to:
 https://github.com/MGross21/mujoco-toolbox/issues.
 
-"""
+"""  # noqa: D205, D400, D415, W291
 
 from .assets import WORLD_ASSETS, glovebox
 from .builder import Builder
@@ -39,7 +39,7 @@ from .controllers import (
 from .utils import _Platform
 from .wrapper import Wrapper
 
-__version__ = "0.4.5"
+__version__ = "5.0.1-rc.1"
 __author__ = "Michael Gross"
 __github_repo__ = "mujoco-toolbox"
 __license__ = "MIT"
@@ -72,8 +72,21 @@ CAPTURE_PARAMETERS = [  # MjData default fields to capture during simulation
     "sensordata",
 ]
 
+class SimulationError(Exception):
+    """Custom exception for simulation-related errors."""
+
+    __doc__ = None  # Exclude from Sphinx documentation
+
+
+class SimulationWarning(Warning):
+    """Custom warning for simulation-related issues."""
+
+    __doc__ = None  # Exclude from Sphinx documentation
+
+
 if __version__.startswith("0"):
     from .utils import _print_warning
+
     _print_warning(
         f"{__package__} (v{__version__}) is still under development.",
         f"Report any issues to https://github.com/MGross21/{__github_repo__}/issues",